pyproject.toml
[project]
name = "callbot-refactor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beeflow",
    "fastapi>=0.121.1",
    "soundfile>=0.13.1",
    "websocket-client>=1.9.0",
    "websockets>=15.0.1",
]

#[tool.uv]
#index-url = "https://registry.aws.platform.vpbank.dev/repository/pypi-proxy/simple"
[tool.uv.sources]
# beeflow = { git = "https://git.aws.platform.vpbank.dev/bic/ds2/beeflow" }
# beeflow = { path = "C:/Users/anhnd98/workspace/beeflow-main/beeflow" }
beeflow = { git = "https://github.com/thanglq150188/beeflow" }
